[P4ADEV-4507] create CieDebtPositionFacadeService, removed dummy debt position creation logic, update README, refactor CIE OpenAPI generator to use YAML OpenAPI
